+++ packages/e2e-cypress/package.json
@@ -34,6 +34,7 @@
     "@cypress/webpack-dev-server": "^1.6.0",
     "cross-env": "^7.0.3",
     "cypress": "^8.0.0",
+    "clone-deep": "^4.0.1",
     "json5": "^2.2.0",
     "start-server-and-test": "^1.12.6"
   },
