+++ packages/e2e-cypress/src/helpers/index.ts
@@ -1,3 +1,4 @@
-import './commands';
+// import './commands';
 export * from './install-quasar-plugin';
 export * from './generate-webpack-config';
+export * from './commands';
