+++ packages/e2e-cypress/src/templates/typescript/test/cypress/support/commands.ts
@@ -26,4 +26,6 @@
 
 // DO NOT REMOVE
 // Imports Quasar Cypress AE predefined commands
-import '@quasar/quasar-app-extension-testing-e2e-cypress';
+import { registerCommands } from '@quasar/quasar-app-extension-testing-e2e-cypress';
+
+registerCommands();
